--- MineX/res/models/player_dim.obj
@@ -1,44 +1,0 @@
-# Blender v2.80 (sub 75) OBJ File: ''
-# www.blender.org
-o Cube_Cube.002
-v -0.250000 -1.000000 0.500000
-v -0.250000 1.000000 0.500000
-v -0.250000 -1.000000 -0.500000
-v -0.250000 1.000000 -0.500000
-v 0.250000 -1.000000 0.500000
-v 0.250000 1.000000 0.500000
-v 0.250000 -1.000000 -0.500000
-v 0.250000 1.000000 -0.500000
-vt 0.625000 0.000000
-vt 0.375000 0.250000
-vt 0.375000 0.000000
-vt 0.625000 0.250000
-vt 0.375000 0.500000
-vt 0.625000 0.500000
-vt 0.375000 0.750000
-vt 0.625000 0.750000
-vt 0.375000 1.000000
-vt 0.125000 0.750000
-vt 0.125000 0.500000
-vt 0.875000 0.500000
-vt 0.625000 1.000000
-vt 0.875000 0.750000
-vn -1.0000 0.0000 0.0000
-vn 0.0000 0.0000 -1.0000
-vn 1.0000 0.0000 0.0000
-vn 0.0000 0.0000 1.0000
-vn 0.0000 -1.0000 0.0000
-vn 0.0000 1.0000 0.0000
-s off
-f 2/1/1 3/2/1 1/3/1
-f 4/4/2 7/5/2 3/2/2
-f 8/6/3 5/7/3 7/5/3
-f 6/8/4 1/9/4 5/7/4
-f 7/5/5 1/10/5 3/11/5
-f 4/12/6 6/8/6 8/6/6
-f 2/1/1 4/4/1 3/2/1
-f 4/4/2 8/6/2 7/5/2
-f 8/6/3 6/8/3 5/7/3
-f 6/8/4 2/13/4 1/9/4
-f 7/5/5 5/7/5 1/10/5
-f 4/12/6 2/14/6 6/8/6
