--- MineX/res/models/player_real.obj
@@ -1,166 +1,0 @@
-# Blender v2.80 (sub 75) OBJ File: 'player.blend'
-# www.blender.org
-o Cube_Cube.001
-v -0.125000 -1.000000 0.250000
-v -0.250000 1.000000 0.250000
-v 0.125000 -1.000000 0.250000
-v 0.250000 1.000000 0.250000
-v 0.125000 -1.000000 -0.250000
-v -0.125000 -1.000000 -0.250000
-v 0.250000 0.999994 -0.250000
-v -0.250000 0.999994 -0.250000
-v -0.125000 0.500000 0.500000
-v -0.125000 0.500000 -0.500000
-v 0.125000 0.500000 -0.500000
-v 0.125000 0.500000 0.500000
-v 0.125000 0.500000 0.250000
-v -0.125000 0.500000 0.250000
-v 0.125000 0.500000 -0.250000
-v -0.125000 0.500000 -0.250000
-v -0.125000 -0.250000 -0.500000
-v 0.125000 -0.250000 -0.500000
-v 0.125000 -0.250000 0.500000
-v 0.125000 -0.250000 0.250000
-v -0.125000 -0.250000 0.250000
-v 0.125000 -0.250000 -0.250000
-v -0.125000 -0.250000 -0.250000
-v -0.125000 -0.250000 0.500000
-v -0.125000 0.500000 0.250000
-v -0.125000 0.500000 -0.250000
-v -0.250000 0.500000 0.250000
-v 0.250000 0.500000 0.250000
-v -0.250000 0.500000 -0.250000
-v 0.250000 0.500000 -0.250000
-vt 0.600000 0.800000
-vt 0.400003 0.600000
-vt 0.400000 0.800000
-vt 0.300000 0.800000
-vt 0.200000 1.000000
-vt 0.200000 0.800000
-vt 0.000000 0.999998
-vt 0.200000 0.800000
-vt 0.000000 0.800000
-vt 0.900000 0.500000
-vt 0.800000 0.800000
-vt 0.800000 0.500000
-vt 0.400000 0.800000
-vt 0.200000 0.600000
-vt 0.200000 0.800000
-vt 0.700000 0.600000
-vt 0.500000 0.300000
-vt 0.700000 0.300000
-vt 0.400000 0.600000
-vt 0.300000 0.300000
-vt 0.400000 0.300000
-vt 0.800000 0.600000
-vt 0.800000 0.300000
-vt 0.300000 0.600000
-vt 0.100000 0.300000
-vt 0.900000 0.500000
-vt 1.000000 0.800000
-vt 0.900000 0.800000
-vt 0.100000 0.600000
-vt 0.000000 0.300000
-vt 0.900000 0.200000
-vt 1.000000 0.500000
-vt 0.900000 0.500000
-vt 0.500000 0.600000
-vt 0.400000 0.300000
-vt 0.300000 0.900000
-vt 0.400000 0.800000
-vt 0.400000 0.900000
-vt 0.000003 0.800000
-vt 0.200000 0.600000
-vt 0.000000 0.600000
-vt 0.500000 0.800000
-vt 0.400000 0.900000
-vt 0.400000 0.800000
-vt 0.600000 0.600000
-vt 0.800000 0.800000
-vt 0.800000 0.600000
-vt 0.100000 0.000000
-vt 0.300000 0.000000
-vt 0.700000 0.900000
-vt 0.600000 0.800000
-vt 0.700000 0.800000
-vt 0.500000 0.000000
-vt 0.700000 0.000000
-vt 0.800000 0.500000
-vt 0.900000 0.200000
-vt 0.900000 0.500000
-vt 0.600000 0.900000
-vt 0.500000 0.800000
-vt 0.600000 0.800000
-vt 1.000000 0.000000
-vt 0.800000 0.200000
-vt 1.000000 0.200000
-vt 0.600000 0.600000
-vt 0.300000 1.000000
-vt 0.200000 0.999998
-vt 0.900000 0.800000
-vt 0.400000 0.600000
-vt 1.000000 0.500000
-vt 0.000000 0.600000
-vt 1.000000 0.200000
-vt 0.400000 0.600000
-vt 0.300000 0.800000
-vt 0.200000 0.800000
-vt 0.500000 0.900000
-vt 0.600000 0.800000
-vt 0.600000 0.900000
-vt 0.800000 0.200000
-vt 0.500000 0.900000
-vt 0.800000 0.000000
-vn 1.0000 0.0000 0.0000
-vn 0.0000 -1.0000 0.0000
-vn 0.0000 0.0000 -1.0000
-vn 0.0000 0.0000 1.0000
-vn -1.0000 0.0000 0.0000
-vn 0.0000 1.0000 0.0000
-s off
-f 28/1/1 7/2/1 4/3/1
-f 5/4/2 1/5/2 6/6/2
-f 30/7/3 8/8/3 7/9/3
-f 22/10/3 6/11/3 23/12/3
-f 27/13/4 4/14/4 2/15/4
-f 14/16/5 23/17/5 21/18/5
-f 11/19/1 22/20/1 18/21/1
-f 9/22/5 21/18/5 24/23/5
-f 15/24/1 20/25/1 22/20/1
-f 12/26/4 24/27/4 19/28/4
-f 13/29/1 19/30/1 20/25/1
-f 10/31/3 18/32/3 17/33/3
-f 16/34/5 17/35/5 23/17/5
-f 10/36/6 15/37/6 11/38/6
-f 8/39/5 27/40/5 2/41/5
-f 12/42/6 25/43/6 9/44/6
-f 28/45/2 29/46/2 30/47/2
-f 22/20/1 3/48/1 5/49/1
-f 23/50/2 18/51/2 22/52/2
-f 21/18/5 6/53/5 1/54/5
-f 3/55/4 21/56/4 1/57/4
-f 19/58/2 21/59/2 20/60/2
-f 7/61/6 2/62/6 4/63/6
-f 28/1/1 30/64/1 7/2/1
-f 5/4/2 3/65/2 1/5/2
-f 30/7/3 29/66/3 8/8/3
-f 22/10/3 5/67/3 6/11/3
-f 27/13/4 28/68/4 4/14/4
-f 14/16/5 16/34/5 23/17/5
-f 11/19/1 15/24/1 22/20/1
-f 9/22/5 14/16/5 21/18/5
-f 15/24/1 13/29/1 20/25/1
-f 12/26/4 9/69/4 24/27/4
-f 13/29/1 12/70/1 19/30/1
-f 10/31/3 11/71/3 18/32/3
-f 16/34/5 10/72/5 17/35/5
-f 10/36/6 26/73/6 15/37/6
-f 8/39/5 29/74/5 27/40/5
-f 12/42/6 13/75/6 25/43/6
-f 28/45/2 27/76/2 29/46/2
-f 22/20/1 20/25/1 3/48/1
-f 23/50/2 17/77/2 18/51/2
-f 21/18/5 23/17/5 6/53/5
-f 3/55/4 20/78/4 21/56/4
-f 19/58/2 24/79/2 21/59/2
-f 7/61/6 8/80/6 2/62/6
