+++ bin/de/keygalp/mineX/fontRendering/fontVertex.txt
@@ -1,0 +1,15 @@
+#version 330
+
+in vec2 position;
+in vec2 textureCoords;
+
+out vec2 pass_textureCoords;
+
+uniform vec2 translation;
+
+void main(){
+
+	gl_Position = vec4(position.x + translation.x+1, position.y+translation.y-1, 0.0, 1.0);
+	pass_textureCoords = textureCoords;
+	//pass_textureCoords =  vec2((gl_Position.x+1.0)/2.0, 1 - (gl_Position.y+1.0)/2.0);
+}
