+++ bin/de/keygalp/mineX/guis/rendering/guiFragmentShader.txt
@@ -1,0 +1,13 @@
+#version 140
+
+in vec2 textureCoords;
+
+out vec4 out_Color;
+
+uniform sampler2D guiTexture;
+
+void main(void){
+
+	out_Color = texture(guiTexture,textureCoords);
+
+}
