+++ bin/de/keygalp/mineX/guis/rendering/guiObjectFragmentShader.txt
@@ -1,0 +1,13 @@
+#version 140
+
+in vec2 textureCoords;
+
+out vec4 out_Color;
+
+uniform vec4 col;
+
+void main(void){
+
+	out_Color = col;
+
+}
