+++ src/de/keygalp/mineX/Game.java
@@ -67,16 +67,21 @@ public class Game extends DisplayManager {
 
 		CommandManager.registerAll();
 
-		TexturedModel playerModel = new TexturedModel(loader.loadToVao(OBJLoader.loadOBJ("player")),
-				new ModelTexture(loader.loadTexture("/textures/playerLayout")));
+		TexturedModel playerModel = new TexturedModel(loader.loadToVao(OBJLoader.loadOBJ("player_real")),
+				new ModelTexture(loader.loadTexture("/textures/PlayerUV")));
 
-		player = new Player(this, playerModel, new Vector3f(0.5f, World.SUPERFLAT_HEIGHT+4, 0.5f), 0, 0, 0, 0.1f);
+		player = new Player(this, playerModel, new Vector3f(0, World.SUPERFLAT_HEIGHT+4, 0), 0, 0, 0, 0.1f);
 
 		guiRenderer = new GUIRenderer(loader);
 
 		hudManager = new HUDManager(this, loader, guiRenderer);
 
 		camera.setPitch(180);
+		
+		item = new ItemEntity(playerModel, new Vector3f(0.5f,World.SUPERFLAT_HEIGHT+2, 0.5f), 0,0,0,1);
+		
+		renderer.processEntity(item);
+		
 
 		lastTick = System.nanoTime();
