+++ src/de/keygalp/mineX/entities/ItemEntity.java
@@ -6,9 +6,10 @@ import de.keygalp.mineX.models.TexturedModel;
 
 public class ItemEntity extends Entity{
 
-	public ItemEntity(TexturedModel model, Vector3f position, float rotX, float rotY, float rotZ, float scale,
-			BoundingBox bounds) {
-		super(model, position, rotX, rotY, rotZ, scale, bounds);
+	static BoundingBox box = new BoundingBox(new Vector3f(1,1,1), new Vector3f(1,1,1), new Vector3f(1,1,1));
+	
+	public ItemEntity(TexturedModel model, Vector3f position, float rotX, float rotY, float rotZ, float scale) {
+		super(model, position, rotX, rotY, rotZ, scale, box);
 	}
 
 }
